chore(migration): clean up migration code a bit
