+++ core/app/c/[communitySlug]/settings/legacy-migration/MigrationForm.tsx
@@ -37,7 +37,7 @@ import { Input } from "ui/input";
 import { FormSubmitButton } from "ui/submit-button";
 import { toast } from "ui/use-toast";
 
-import type { getToBeDeletedStructure } from "~/lib/server/legacy-migration/legacy-migration";
+import type { getToBeDeletedStructure } from "~/lib/server/legacy-migration/cleanup";
 import { DataTable } from "~/app/components/DataTable/v2/DataTable";
 import { getPubTitle } from "~/lib/pubs";
 import { didSucceed, useServerAction } from "~/lib/serverActions";
